fix theme on anon
