+++ apps/dotcom/client/src/tla/components/TlaButton/TlaButton.tsx
@@ -10,7 +10,7 @@ export const TlaButton = forwardRef<
 		icon?: string
 		iconRight?: string
 		ghost?: boolean
-		variant?: 'primary' | 'secondary' | 'warning'
+		variant?: 'primary' | 'secondary' | 'warning' | 'cta'
 	}
 >(function TlaButton(
 	{
@@ -33,9 +33,9 @@ export const TlaButton = forwardRef<
 			ref={ref}
 			data-state={isLoading ? 'loading' : 'ready'}
 			className={classNames(
-				'tla-text_ui__medium',
 				styles.button,
 				{
+					[styles.cta]: variant === 'cta',
 					[styles.primary]: variant === 'primary',
 					[styles.secondary]: variant === 'secondary',
 					[styles.ghost]: ghost,
