+++ packages/tldraw/src/lib/ui/components/LanguageMenu.tsx
@@ -1,4 +1,4 @@
-import { LANGUAGES, useEditor, useValue } from '@tldraw/editor'
+import { LANGUAGES, useMaybeEditor, useValue } from '@tldraw/editor'
 import { useUiEvents } from '../context/events'
 import { TldrawUiMenuCheckboxItem } from './primitives/menus/TldrawUiMenuCheckboxItem'
 import { TldrawUiMenuGroup } from './primitives/menus/TldrawUiMenuGroup'
@@ -6,9 +6,11 @@ import { TldrawUiMenuSubmenu } from './primitives/menus/TldrawUiMenuSubmenu'
 
 /** @public @react */
 export function LanguageMenu() {
-	const editor = useEditor()
+	const editor = useMaybeEditor()
 	const trackEvent = useUiEvents()
-	const currentLanguage = useValue('locale', () => editor.user.getLocale(), [editor])
+	const currentLanguage = useValue('locale', () => editor?.user.getLocale(), [editor])
+
+	if (!editor) return null
 
 	return (
 		<TldrawUiMenuSubmenu id="help menu language" label="menu.language">
